Restore the test for #2202 (concurrent use of the linker).
The previous test had basically been inhibited when we switched
the default environment from Rhino to Node.js. Indeed, this caused
`loadedJSEnv` not to use the linker itself, but rather depend on
the result of `fastOptJS`. The test was therefore not testing
anything anymore.

We restore this test in a more explicit and robust way, with a
custom task to concurrently use the linker of `fullOptJS`, and a
dedicated test task to directly depend on the former + `fullOptJS`.
We use `fullOptJS` instead of `fastOptJS` because it takes more
time, increasing the likelihood of concurrent execution.